% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AP.BottledWater.r
\docType{data}
\name{AP.BottledWater}
\alias{AP.BottledWater}
\title{Taste scores of 31 bottled water and tap water samples}
\format{
A data frame with 31 rows and 2 variables:
\describe{
  \item{Tap}{Score for tap water}
  \item{Bottled}{Score for bottled water}
  ...
}
}
\source{
\url{http://www.isi-stats.com/isi/data/chap7/AP/BottledWater.txt}
}
\usage{
AP.BottledWater
}
\description{
A dataset containing the scores of 31 subjects rating of bottled and tap water
}
\keyword{datasets}
